Regular App maintenance
Introduced lazy loading and CSS improvements
