+++ src/Components/LoadingIndicator/LoadingIndicator.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Styles from "./LoadingIndicator.module.css";
+
+const LoadingIndicator = () => (
+  <div className={Styles.loadingIndicator}>Loading...</div>
+);
+
+export default LoadingIndicator;
