+++ src/Components/Resume/AboutMe/AboutMe.module.css
@@ -7,10 +7,9 @@
 }
 
 .aboutme h1 {
-  font-size: 1.5rem;
+  font-size: 2.625rem;
   font-weight: bolder;
   color: #333;
-  letter-spacing: 0.4rem;
   margin-left: auto;
   margin-right: auto;
   margin-top: 4vh;
@@ -18,8 +17,7 @@
 }
 
 .aboutme p {
-  font-size: 1rem;
-  letter-spacing: 0.05rem;
+  font-size: 1.25rem;
   line-height: 1.5;
   margin: 4vh 4vw;
 }
