+++ src/Components/Resume/Resume.js
@@ -1,12 +1,13 @@
-import React from "react";
+import React, { Suspense } from "react";
 import Header from "./Header/Header";
 import AboutMe from "./AboutMe/index";
 import Portfolio from "./Portfolio/index";
 import Vita from "./Vita/index";
-import { Parallax } from "react-parallax";
-import Wallpaper1 from "../../Images/wallpaper1.jpg";
 import Styles from "./Resume.module.css";
 import Footer from "./Footer/index";
+import LoadingIndicator from "../LoadingIndicator/LoadingIndicator";
+
+const ParallaxElement = React.lazy(() => import("./ParallaxContainer"));
 
 const personalInfo = {
   name: "Harish Baskar",
@@ -17,22 +18,15 @@ const personalInfo = {
   linkedinurl: "https://www.linkedin.com/in/harish-baskar-96589a80",
 };
 
-const transformElement = (percentage) => {
-  return (
-    <div>
-      <div
-        style={{
-          position: "absolute",
-          background: `rgba(75, 67, 219, ${percentage * 0.5})`,
-          transform: "scale(4, 3)",
-          borderRadius: "50%",
-          width: percentage * 700,
-          height: percentage * 700,
-        }}
-      />
-    </div>
-  );
-};
+const ParallaxItem = () => (
+  <div>
+    <Suspense fallback={<LoadingIndicator />}>
+      <section>
+        <ParallaxElement />
+      </section>
+    </Suspense>
+  </div>
+);
 
 const resume = () => {
   return (
@@ -42,40 +36,13 @@ const resume = () => {
         <div className={Styles.whitesection}>
           <AboutMe />
         </div>
-        <Parallax
-          bgImage={Wallpaper1}
-          strength={300}
-          className={Styles.parallaxcontainer}
-          renderLayer={(percentage) => transformElement(percentage)}
-        >
-          <div style={{ height: 700 }}>
-            <div className={Styles.insidestyles}></div>
-          </div>
-        </Parallax>
+        <ParallaxItem />
         <Vita />
-        <Parallax
-          bgImage={Wallpaper1}
-          strength={400}
-          className={Styles.parallaxcontainer}
-          renderLayer={(percentage) => transformElement(percentage)}
-        >
-          <div style={{ height: 800 }}>
-            <div className={Styles.insidestyles}></div>
-          </div>
-        </Parallax>
+        <ParallaxItem />
         <div className={Styles.whitesection}>
           <Portfolio />
         </div>
-        <Parallax
-          bgImage={Wallpaper1}
-          strength={400}
-          className={Styles.parallaxcontainer}
-          renderLayer={(percentage) => transformElement(percentage)}
-        >
-          <div style={{ height: 800 }}>
-            <div className={Styles.insidestyles}></div>
-          </div>
-        </Parallax>
+        <ParallaxItem />
         <div className={Styles.whitesection}>
           <Footer />
         </div>
