+++ src/Pages/Projects/Projects.module.css
@@ -11,14 +11,13 @@
   display: flex;
   flex-direction: column;
   background-color: transparent;
-  letter-spacing: 0.1rem;
   margin-left: 1rem;
   margin-right: 1rem;
   padding-bottom: 2rem;
 }
 
 .card {
-  background: rgb(255, 250, 230, 0.9);
+  background: #ffffff;
   font-size: 2rem;
   max-width: 100vw;
   margin: 2rem;
@@ -38,7 +37,7 @@
 }
 
 .topcontainer h1 {
-  font-size: 1.5rem;
+  font-size: 2.625rem;
   font-weight: bolder;
   color: #ffe479;
   margin-left: auto;
@@ -48,10 +47,9 @@
 }
 
 .projects h2 {
-  font-size: 1.5rem;
+  font-size: 2rem;
   font-weight: bolder;
   color: #555;
-  letter-spacing: 0.25rem;
   margin-left: auto;
   margin-right: auto;
   margin-bottom: 2vh;
@@ -62,14 +60,12 @@
   color: #555;
   margin-bottom: 1rem;
   font-size: 1rem;
-  letter-spacing: 0.15rem;
 }
 
 .projects p {
   color: #555;
   margin-bottom: 1rem;
   font-size: 1rem;
-  letter-spacing: 0.15rem;
 }
 
 .projects li {
