+++ src/index.css
@@ -82,7 +82,6 @@ html {
   -webkit-font-smoothing: antialiased;
   line-height: 1.5;
   min-height: 100vh;
-  text-shadow: 0 3px 5px rgba(0, 0, 0, 0.1);
 }
 
 a {
